Adds lost <b> tag

+++ code/game/objects/structures/stool_bed_chair_nest/wheelchair.dm
@@ -11,6 +11,12 @@
 	var/driving = FALSE
 	var/bloodiness
 
+/obj/structure/bed/chair/wheelchair/Initialize()
+	. = ..()
+
+	if(!item_form_type)
+		verbs -= .verb/collapse
+
 /obj/structure/bed/chair/wheelchair/on_update_icon()
 	return
 
@@ -119,6 +125,9 @@
 	set category = "Object"
 	set src in oview(1)
 
+	if(!item_form_type)
+		return
+
 	if(!CanPhysicallyInteract(usr))
 		return
 
@@ -129,10 +138,10 @@
 		return
 
 	if(buckled_mob)
-		to_chat(usr, SPAN_WARNING("You can't collapse \the [src.name] while it still on use."))
+		to_chat(usr, SPAN_WARNING("You can't collapse \the [src.name] while it is still in use."))
 		return
 
-	usr.visible_message("<b>[usr]</b> starts collapse \the [src.name].")
+	usr.visible_message("<b>[usr]</b> starts to collapse \the [src.name].")
 	if(do_after(usr, 4 SECONDS, src))
 		var/obj/item/wheelchair_kit/K = new item_form_type(get_turf(src))
 		visible_message(SPAN_NOTICE("<b>[usr]</b> collapses \the [src.name]."))
@@ -150,9 +159,12 @@
 	var/structure_form_type = /obj/structure/bed/chair/wheelchair
 
 /obj/item/wheelchair_kit/attack_self(mob/user)
-	user.visible_message("<b>[user]</b> starts lay out \the [src.name].")
+	if(!structure_form_type)
+		return
+
+	user.visible_message("<b>[user]</b> starts to lay out \the [src.name].")
 	if(do_after(user, 4 SECONDS, src))
 		var/obj/structure/bed/chair/wheelchair/W = new structure_form_type(get_turf(user))
-		user.visible_message(SPAN_NOTICE("[user]</b> lay out \the [W.name]."))
+		user.visible_message(SPAN_NOTICE("<b>[user]</b> lays out \the [W.name]."))
 		W.add_fingerprint(user)
 		qdel(src)
